gymcare/requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
billiard==4.2.1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.43.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
Django==5.0.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-fcm==0.1.1
django-filter==25.1
django-flatpickr==2.0.3
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
dnspython==2.7.0
dotenv==0.9.9
drf-yasg==1.21.10
email_validator==2.2.0
fcm-django==2.2.1
firebase-admin==6.8.0
flatbuffers==25.2.10
frozenlist==1.6.0
gast==0.6.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
joblib==1.5.1
jwcrypto==1.5.6
keras==3.10.0
kombu==5.5.0
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
multidict==6.4.4
namex==0.1.0
numpy==2.1.3
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyfcm==2.0.8
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
